requirements.txt
PySide6
loguru
SQLAlchemy
aiosqlite
PyYAML